refactor: ponytail leftover cuts — shared panel + document joins
- Unify LinkForgottenExpensesPanel (invoice/job modes) and wire both pages
- Parameterize documentJoins / documentLocationSelect; keep deprecated aliases
- Unit tests for join/select SQL shape
- Skip log subpath deletion: main now imports @ai-fsm/log/web
